# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Automation Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all workflows
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a new workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Created workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get a workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowBody"
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workflows/{id}/execute:
    post:
      operationId: executeWorkflow
      tags: [workflows]
      summary: Execute a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                inputData:
                  type: object
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /workflows/{id}/toggle:
    post:
      operationId: toggleWorkflow
      tags: [workflows]
      summary: Toggle workflow active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    WorkflowNode:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        label:
          type: string
        x:
          type: number
        y:
          type: number
        config:
          type: object
      required: [id, type, label, x, y]
    WorkflowEdge:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        sourceHandle:
          type: string
        targetHandle:
          type: string
      required: [id, source, target]
    Workflow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        webhookToken:
          type: string
          nullable: true
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdge"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        lastRunStatus:
          type: string
          nullable: true
        totalRuns:
          type: integer
        successRuns:
          type: integer
        errorRuns:
          type: integer
      required: [id, name, active, nodes, edges, createdAt, updatedAt, totalRuns, successRuns, errorRuns]
    CreateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdge"
      required: [name]
    UpdateWorkflowBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdge"
    Execution:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        workflowName:
          type: string
        status:
          type: string
          enum: [success, error, running]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        nodeResults:
          type: array
          items:
            $ref: "#/components/schemas/NodeResult"
        inputData:
          type: object
        outputData:
          type: object
          nullable: true
      required: [id, workflowId, workflowName, status, startedAt, nodeResults, inputData]
    NodeResult:
      type: object
      properties:
        nodeId:
          type: string
        nodeType:
          type: string
        status:
          type: string
          enum: [success, error, skipped]
        output:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        durationMs:
          type: integer
      required: [nodeId, nodeType, status, durationMs]
...